+++ src/routes/setting/vendorConfig/getVendorList.ts
@@ -4,20 +4,20 @@ import u from "@/utils";
 const router = express.Router();
 
 export default router.post("/", async (req, res) => {
-  const data = await u.db("o_vendorConfig").select("*");
+  const data = await u.db("o_vendorConfig").where("userId", (req as any).user.id).select("*");
 
   const list = (
     await Promise.all(
       data.map(async (item) => {
         const vendor = u.vendor.getVendor(item.id!);
         if (!vendor) {
-          await u.db("o_vendorConfig").where("id", item.id).delete();
+          await u.db("o_vendorConfig").where("id", item.id).andWhere("userId", (req as any).user.id).delete();
           return null
         };
         return {
           ...item,
           inputValues: JSON.parse(item.inputValues ?? "{}"),
-          models: await u.vendor.getModelList(item.id!),
+          models: await u.vendor.getModelList(item.id!, (req as any).user.id),
           code: u.vendor.getCode(item.id!),
           description: vendor.description ?? "",
           inputs: vendor.inputs,
